+++ Pandora/Paradigm/Structure/Variation/Substructure.hs
@@ -1,7 +1,8 @@
 module Pandora.Paradigm.Structure.Variation.Substructure (Substructure (..)) where
 
 import Pandora.Paradigm.Inventory.Optics (type (:-.))
+import Pandora.Paradigm.Primary.Functor.Tagged (Tagged)
 
 class Substructure f t where
-	type Output (f :: * -> k) (t :: * -> *) a = r | r -> f t a
-	sub :: t a :-. Output f t a
+	type Output (f :: * -> k) (t :: * -> *) a
+	sub :: t a :-. Tagged f (Output f t a)